+++ apps/wallet/instance/src/main/java/com/tonapps/tonkeeper/extensions/FlowExt.kt
@@ -5,21 +5,22 @@ import com.tonapps.wallet.data.rates.entity.RatesEntity
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.filterNotNull
+import java.math.BigDecimal
 
 fun formattedRate(
     rateFlow: Flow<RatesEntity>,
-    amountFlow: Flow<Float>,
+    amountFlow: Flow<BigDecimal>,
     token: String,
-    formatter: (currencyCode: String, amount: Float) -> CharSequence = CurrencyFormatter::format
+    formatter: (currencyCode: String, amount: BigDecimal) -> CharSequence = CurrencyFormatter::format
 ): Flow<CharSequence> = combine(rateFlow, amountFlow) { rates, amount ->
     formatRate(rates, amount, token, formatter)
 }.filterNotNull()
 
 fun formatRate(
     rates: RatesEntity,
-    amount: Float,
+    amount: BigDecimal,
     token: String,
-    formatter: (currencyCode: String, amount: Float) -> CharSequence = CurrencyFormatter::format
+    formatter: (currencyCode: String, amount: BigDecimal) -> CharSequence = CurrencyFormatter::format
 ): CharSequence? {
     val rate = rates.rate(token) ?: return null
     val totalAmount = rate.value * amount
